third_party/WebKit/LayoutTests/fast/storage/disallowed-storage.html

<!DOCTYPE html>
<html>
<head>
</head>
<body>
    <script src="../js/resources/js-test-pre.js"></script>
    <script>
        description("This tests that storage methods throw exceptions when storage is disabled with reasonable messages.");

        if (window.testRunner)
            testRunner.setStorageAllowed(false);

        shouldThrow('window.sessionStorage');
        shouldThrow('window.localStorage');
    </script>
    <script src="../js/resources/js-test-post.js"></script>
</body>
</html>
