third_party/WebKit/LayoutTests/fast/storage/disallowed-storage-expected.txt

This tests that storage methods throw exceptions when storage is disabled with reasonable messages.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.sessionStorage threw exception SecurityError: Access to 'sessionStorage' is denied for this document..
PASS window.localStorage threw exception SecurityError: Access to 'localStorage' is denied for this document..
PASS successfullyParsed is true

TEST COMPLETE
